JPA sprint 3

+++ demo/src/main/java/com/example/demo/servicio/ClienteService.java
@@ -9,13 +9,13 @@ import com.example.demo.entidad.Cliente;
 @Service
 public interface ClienteService {
 
-    public Cliente findById(int id);
+    public Cliente findById(Long id);
 
     public Collection<Cliente> findAll();
 
     public void add(Cliente cliente);
 
-    public void deleteById(int id);
+    public void deleteById(Long id);
 
     public void update(Cliente cliente);
 }
